langchain/email-sms-tools/requirements.txt
commune-mail
langchain
langchain-openai
openai
